requirements.txt
Flask
streamlit
python-dotenv
langchain
langchain-core
langchain-groq
requests
flask-cors
gunicon